feat(common): restructure, integration tests and logic fix for primed input processing
