+++ src/plugins/common/src/tools/action_key/user_input.rs
@@ -3,10 +3,17 @@ use crate::traits::handles_localization::Token;
 use bevy::prelude::*;
 use serde::{Deserialize, Serialize};
 
+#[derive(Debug, PartialEq, Eq, Clone, Hash, SystemSet)]
+pub struct InputSystem;
+
 #[derive(TypePath, Debug, PartialEq, Eq, Hash, Clone, Copy, Serialize, Deserialize)]
 pub enum UserInput {
 	KeyCode(KeyCode),
 	MouseButton(MouseButton),
+}
+
+impl UserInput {
+	pub const SYSTEM: InputSystem = InputSystem;
 }
 
 impl From<UserInput> for Token {
